requirements.txt
ollama
faster-whisper
pyttsx3
rich
pydantic-settings
SpeechRecognition
pyaudio
pywin32
python-dotenv
aiohttp
pywebview
fastapi
uvicorn
jinja2
psutil
screen_brightness_control
pycaw
pygetwindow
keyboard
